refactor: sidebar v2
