+++ src/components/SideBar/SideBar.jsx
@@ -1,71 +1,38 @@
-import { useState } from "react";
+import { FaRegCalendarAlt, FaUserTie } from "react-icons/fa";
+import { FiHome } from "react-icons/fi";
+import { GiReceiveMoney } from "react-icons/gi";
+import { GrUserManager } from "react-icons/gr";
+import { Link } from "react-router-dom";
+import logoImg from "/logoGesco2.png";
+import LinkSideBar from "./LinkSideBar/LinkSideBar";
 
 export function SideBar() {
-	const [isOpen, setIsOpen] = useState(false);
-
-	const toggleSidebar = () => {
-		setIsOpen(!isOpen);
-	};
-
 	return (
-		<div className="flex  max-w-fit">
-			{/* Botão de menu para mobile */}
-			<button
-				onClick={toggleSidebar}
-				className="md:hidden p-4 text-white bg-blue-900"
-				type="button"
-			></button>
-
-			{/* Sidebar */}
-			<div
-				className={`${
-					isOpen ? "translate-x-0" : "-translate-x-full"
-				} md:translate-x-0 fixed inset-y-0 left-0 w-64 bg-blue-900 text-white transform transition-transform duration-300 ease-in-out`}
+		<nav className="lg:min-w-full max-w-10 min-h-screen bg-blue-900  flex flex-col items-center gap-7 overflow-visible relative">
+			<Link
+				to="/"
+				className="flex items-center justify-center gap-2 p-3 no-underline bg-orange-600 absolute top-0 left-0 w-[100%]"
 			>
-				<div className="flex items-center justify-center h-20 bg-orange-600">
-					<span className="text-2xl font-semibold">College</span>
-				</div>
-				<nav className="flex flex-col mt-10 space-y-6">
-					<a
-						href="#dashboard"
-						className="flex items-center px-6 py-3 text-lg hover:bg-blue-800"
-					>
-						Dashboard
-					</a>
-					<a
-						href="#alunos"
-						className="flex items-center px-6 py-3 text-lg hover:bg-blue-800"
-					>
-						Alunos
-					</a>
-					<a
-						href="#docentes"
-						className="flex items-center px-6 py-3 text-lg hover:bg-blue-800"
-					>
-						Docentes
-					</a>
-					<a
-						href="#eventos"
-						className="flex items-center px-6 py-3 text-lg hover:bg-blue-800"
-					>
-						Eventos
-					</a>
-					<a
-						href="#financas"
-						className="flex items-center px-6 py-3 text-lg hover:bg-blue-800"
-					>
-						Finanças
-					</a>
-				</nav>
-			</div>
-
-			{/* Overlay para quando a sidebar estiver aberta no mobile */}
-			{isOpen && (
-				<div
-					className="fixed inset-0 bg-black opacity-50 z-10 md:hidden"
-					onClick={toggleSidebar}
+				<img src={logoImg} alt="Logo" className="w-8 " />
+				<h2 className="text-2xl text-white font-bold drop-shadow-[0px_0px_3px_rgba(0,0,0,0.75)] hidden sm:block">
+					Gesco
+				</h2>
+			</Link>
+			<ul className="flex flex-col gap-3 w-full pt-20">
+				<LinkSideBar url="/dashboard" icone={<FiHome />} nome="Dashboard" />
+				<LinkSideBar url="/students" icone={<GrUserManager />} nome="Alunos" />
+				<LinkSideBar url="/teachers" icone={<FaUserTie />} nome="Docentes" />
+				<LinkSideBar
+					url="/events"
+					icone={<FaRegCalendarAlt />}
+					nome="Eventos"
 				/>
-			)}
-		</div>
+				<LinkSideBar
+					url="/finance"
+					icone={<GiReceiveMoney />}
+					nome="Finanças"
+				/>
+			</ul>
+		</nav>
 	);
 }
